Re-structure for evaluation callback (Separation between evaluation callback and evaluation policy in Sinergym code)
